Seguranca

+++ src/main/java/learn/spring/api/models/ProductModel.java
@@ -1,16 +1,15 @@
 package learn.spring.api.models;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import org.springframework.hateoas.RepresentationModel;
 
-import jakarta.persistence.Table;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "TB_PRODUCTS")
@@ -22,7 +21,9 @@ public class ProductModel extends RepresentationModel<ProductModel> implements S
     private UUID idProduct;
 
     private String name;
-    private BigDecimal value;
+    
+    //@Column(precision = 6, scale = 4) Corno véi do JPA não tava conseguindo fazer a conversão com BigDecimal, então eu facilitei pra ele(No final, não deu certo)
+    private Double value;
     
     public UUID getIdProduct() {
         return idProduct;
@@ -36,10 +37,10 @@ public class ProductModel extends RepresentationModel<ProductModel> implements S
     public void setName(String name) {
         this.name = name;
     }
-    public BigDecimal getValue() {
+    public Double getValue() {
         return value;
     }
-    public void setValue(BigDecimal value) {
+    public void setValue(Double value) {
         this.value = value;
     }
